313 - Update bug squash
Uplifted the snackbar to the main component of the app.
This allows us to bring snackbar messages when the Sign up component
works / does not work as expected.

Update height of the map in selected category / maintenance plan
Added map height to match the height of the card component. In the
small view mode the image is removed since the height is not present.